+++ lib/rdf/model/statement.ex
@@ -287,6 +287,14 @@ defmodule RDF.Statement do
   def has_bnode?({_, _, _, _} = quad), do: Quad.has_bnode?(quad)
   def has_bnode?({_, _, _} = triple), do: Triple.has_bnode?(triple)
 
+  @doc """
+  Returns a list of all `RDF.BlankNode`s within the given `statement`.
+  """
+  @spec bnodes(t) :: list(BlankNode.t())
+  def bnodes(statement)
+  def bnodes({_, _, _, _} = quad), do: Quad.bnodes(quad)
+  def bnodes({_, _, _} = triple), do: Triple.bnodes(triple)
+
   def include_value?({_, _, _, _} = quad, value), do: Quad.include_value?(quad, value)
   def include_value?({_, _, _} = triple, value), do: Triple.include_value?(triple, value)
 end